Split the TypeScript code in `api/` into multiple modules
