+++ api/src/crud.ts
@@ -1,0 +1,424 @@
+import { mergeDeep, opAsync, responseFromJson } from "./utils.ts";
+import { ChiselCursor, ChiselEntity, requestContext } from "./datastore.ts";
+
+// TODO: BEGIN: when module import is fixed:
+//     import { parse as regExParamParse } from "regexparam";
+// or:
+//     import { parse as regExParamParse } from "regexparam";
+// In the meantime, the regExParamParse function is copied from
+// https://deno.land/x/regexparam@v2.0.0/src/index.js under MIT License included
+// below. ChiselStrike added the TS signature and minor cleanups.
+//
+// Copyright (c) Luke Edwards <luke.edwards05@gmail.com> (lukeed.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+function regExParamParse(str: string, loose: boolean) {
+    let tmp, pattern = "";
+    const keys = [], arr = str.split("/");
+    arr[0] || arr.shift();
+
+    while ((tmp = arr.shift())) {
+        const c = tmp[0];
+        if (c === "*") {
+            keys.push("wild");
+            pattern += "/(.*)";
+        } else if (c === ":") {
+            const o = tmp.indexOf("?", 1);
+            const ext = tmp.indexOf(".", 1);
+            keys.push(tmp.substring(1, ~o ? o : ~ext ? ext : tmp.length));
+            pattern += !!~o && !~ext ? "(?:/([^/]+?))?" : "/([^/]+?)";
+            if (~ext) pattern += (~o ? "?" : "") + "\\" + tmp.substring(ext);
+        } else {
+            pattern += "/" + tmp;
+        }
+    }
+
+    return {
+        keys: keys,
+        pattern: new RegExp("^" + pattern + (loose ? "(?=$|/)" : "/?$"), "i"),
+    };
+}
+// TODO: END: when module import is fixed
+
+type ChiselEntityClass<T extends ChiselEntity> = {
+    new (): T;
+    findOne: (_: { id: string }) => Promise<T | undefined>;
+    findMany: (_: Partial<T>) => Promise<T[]>;
+    build: (...properties: Record<string, unknown>[]) => T;
+    delete: (restrictions: Partial<T>) => Promise<void>;
+    cursor: () => ChiselCursor<T>;
+};
+
+type GenericChiselEntityClass = ChiselEntityClass<ChiselEntity>;
+
+/**
+ * Creates a path parser from a template using regexparam.
+ *
+ * @param pathTemplate the path template such as `/static`, `/param/:id/:otherParam`...
+ * @param loose if true, it can match longer paths. False by default
+ * @returns function that can parse paths given as string.
+ * @see https://deno.land/x/regexparam@v2.0.0
+ */
+export function createPathParser<T extends Record<string, unknown>>(
+    pathTemplate: string,
+    loose = false,
+): (path: string) => T {
+    const { pattern, keys: keysOrFalse } = regExParamParse(pathTemplate, loose);
+    if (typeof keysOrFalse === "boolean") {
+        throw new Error(
+            `invalid pathTemplate=${pathTemplate}, expected string`,
+        );
+    }
+    const keys = keysOrFalse;
+    return function pathParser(path: string): T {
+        const matches = pattern.exec(path);
+        return keys.reduce(
+            (acc: Record<string, unknown>, key: string, index: number) => {
+                acc[key] = matches?.[index + 1];
+                return acc;
+            },
+            {},
+        ) as T;
+    };
+}
+
+/**
+ * Creates a path parser from a template using regexparam.
+ *
+ * @param pathTemplate the path template such as `/static`, `/param/:id/:otherParam`...
+ * @param loose if true, it can match longer paths. False by default
+ * @returns function that can parse paths given in URL.pathname.
+ * @see https://deno.land/x/regexparam@v2.0.0
+ */
+export function createURLPathParser<T extends Record<string, unknown>>(
+    pathTemplate: string,
+    loose = false,
+): (url: URL) => T {
+    const pathParser = createPathParser<T>(pathTemplate, loose);
+    return (url: URL): T => pathParser(url.pathname);
+}
+
+/** Creates a Response object from response body and status. */
+export type CRUDCreateResponse = (
+    body: unknown,
+    status: number,
+) => Promise<Response> | Response;
+
+export type CRUDBaseParams = {
+    /** identifier of the object being manipulated, if any */
+    id?: string;
+    /** ChiselStrike's version/branch the server is running,
+     * such as 'dev' for endpoint '/dev/example'
+     * when using 'chisel apply --version dev'
+     */
+    chiselVersion: string;
+};
+
+export type CRUDMethodSignature<
+    T extends ChiselEntity,
+    E extends ChiselEntityClass<T>,
+    P extends CRUDBaseParams = CRUDBaseParams,
+> = (
+    entity: E,
+    req: Request,
+    params: P,
+    url: URL,
+    createResponse: CRUDCreateResponse,
+) => Promise<Response>;
+
+/**
+ * A dictionary mapping HTTP methods into corresponding REST methods that process a Request and return a Response.
+ */
+export type CRUDMethods<
+    T extends ChiselEntity,
+    E extends ChiselEntityClass<T>,
+    P extends CRUDBaseParams = CRUDBaseParams,
+> = {
+    GET: CRUDMethodSignature<T, E, P>;
+    POST: CRUDMethodSignature<T, E, P>;
+    PUT: CRUDMethodSignature<T, E, P>;
+    PATCH: CRUDMethodSignature<T, E, P>;
+    DELETE: CRUDMethodSignature<T, E, P>;
+};
+
+export type CRUDCreateResponses<
+    T extends ChiselEntity,
+    E extends ChiselEntityClass<T>,
+    P extends CRUDBaseParams = CRUDBaseParams,
+> = {
+    [K in keyof CRUDMethods<T, E, P>]: CRUDCreateResponse;
+};
+
+/**
+ * Fetches crud data based on crud `url`.
+ */
+async function fetchEntitiesCrud<T extends ChiselEntity>(
+    type: { new (): T },
+    url: string,
+): Promise<T[]> {
+    const results = await opAsync(
+        "op_chisel_crud_query",
+        {
+            typeName: type.name,
+            url,
+        },
+        requestContext,
+    );
+    return results as T[];
+}
+
+async function deleteEntitiesCrud<T extends ChiselEntity>(
+    type: { new (): T },
+    url: string,
+): Promise<void> {
+    await opAsync(
+        "op_chisel_crud_delete",
+        {
+            typeName: type.name,
+            url,
+        },
+        requestContext,
+    );
+}
+
+const defaultCrudMethods: CRUDMethods<ChiselEntity, GenericChiselEntityClass> =
+    {
+        // Returns a specific entity matching params.id (if present) or all entities matching the filter in the `filter` URL parameter.
+        GET: async (
+            entity: GenericChiselEntityClass,
+            _req: Request,
+            params: CRUDBaseParams,
+            url: URL,
+            createResponse: CRUDCreateResponse,
+        ) => {
+            const { id } = params;
+            if (id) {
+                const u = await entity.findOne({ id });
+                return createResponse(u ?? "Not found", u ? 200 : 404);
+            } else {
+                return createResponse(
+                    await fetchEntitiesCrud(entity, url.href),
+                    200,
+                );
+            }
+        },
+        // Creates and returns a new entity from the `req` payload. Ignores the payload's id property and assigns a fresh one.
+        POST: async (
+            entity: GenericChiselEntityClass,
+            req: Request,
+            _params: CRUDBaseParams,
+            _url: URL,
+            createResponse: CRUDCreateResponse,
+        ) => {
+            const u = entity.build(await req.json());
+            u.id = undefined;
+            await u.save();
+            return createResponse(u, 200);
+        },
+        // Updates and returns the entity matching params.id (which must be set) from the `req` payload.
+        PUT: async (
+            entity: GenericChiselEntityClass,
+            req: Request,
+            params: CRUDBaseParams,
+            _url: URL,
+            createResponse: CRUDCreateResponse,
+        ) => {
+            const { id } = params;
+            if (!id) {
+                return createResponse(
+                    "PUT requires item ID in the URL",
+                    400,
+                );
+            }
+            const u = entity.build(await req.json());
+            u.id = id;
+            await u.save();
+            return createResponse(u, 200);
+        },
+        PATCH: async (
+            entity: GenericChiselEntityClass,
+            req: Request,
+            params: CRUDBaseParams,
+            _url: URL,
+            createResponse: CRUDCreateResponse,
+        ) => {
+            const { id } = params;
+            if (!id) {
+                return createResponse(
+                    "PATCH requires item ID in the URL",
+                    400,
+                );
+            }
+            const orig = await entity.findOne({ id });
+            if (!orig) {
+                return createResponse(
+                    "object does not exist, cannot PATCH",
+                    404,
+                );
+            }
+            mergeDeep(
+                orig as unknown as Record<string, unknown>,
+                await req.json(),
+            );
+            await orig.save();
+            return createResponse(orig, 200);
+        },
+
+        // Deletes the entity matching params.id (if present) or all entities matching the filter in the `filter` URL parameter. One of the two must be present.
+        DELETE: async (
+            entity: GenericChiselEntityClass,
+            _req: Request,
+            params: CRUDBaseParams,
+            url: URL,
+            createResponse: CRUDCreateResponse,
+        ) => {
+            const { id } = params;
+            if (id) {
+                await entity.delete({ id });
+                return createResponse(`Deleted ID ${id}`, 200);
+            } else {
+                await deleteEntitiesCrud(entity, url.href);
+                return createResponse(
+                    `Deleted entities matching ${url.search}`,
+                    200,
+                );
+            }
+        },
+    } as const;
+
+/**
+ * These methods can be used as `customMethods` in `ChiselStrike.crud()`.
+ *
+ * @example
+ * Put this in the file 'endpoints/comments.ts':
+ * ```typescript
+ * import { Comment } from "../models/comment";
+ * export default crud(
+ *   Comment,
+ *   ':id',
+ *   {
+ *     PUT: standardCRUDMethods.notFound, // do not update, instead returns 404
+ *     DELETE: standardCRUDMethods.methodNotAllowed, // do not delete, instead returns 405
+ *   },
+ * );
+ * ```
+ */
+export const standardCRUDMethods = {
+    forbidden: (
+        _entity: GenericChiselEntityClass,
+        _req: Request,
+        _params: CRUDBaseParams,
+        _url: URL,
+        createResponse: CRUDCreateResponse,
+    ) => Promise.resolve(createResponse("Forbidden", 403)),
+    notFound: (
+        _entity: GenericChiselEntityClass,
+        _req: Request,
+        _params: CRUDBaseParams,
+        _url: URL,
+        createResponse: CRUDCreateResponse,
+    ) => Promise.resolve(createResponse("Not Found", 404)),
+    methodNotAllowed: (
+        _entity: GenericChiselEntityClass,
+        _req: Request,
+        _params: CRUDBaseParams,
+        _url: URL,
+        createResponse: CRUDCreateResponse,
+    ) => Promise.resolve(createResponse("Method Not Allowed", 405)),
+} as const;
+
+/**
+ * Generates endpoint code to handle REST methods GET/PUT/POST/DELETE for this entity.
+ * @example
+ * Put this in the file 'endpoints/comments.ts':
+ * ```typescript
+ * import { Comment } from "../models/comment";
+ * export default crud(Comment, ":id");
+ * ```
+ * This results in a /comments endpoint that correctly handles all REST methods over Comment.
+ * @param entity Entity type
+ * @param urlTemplateSuffix A suffix to be added to the Request URL (see https://deno.land/x/regexparam for syntax).
+ *   Some CRUD methods rely on parts of the URL to identify the resource to apply to. Eg, GET /comments/1234
+ *   returns the comment entity with id=1234, while GET /comments returns all comments. This parameter describes
+ *   how to find the relevant parts in the URL. Default CRUD methods (see `defaultCrudMethods`) look for the :id
+ *   part in this template to identify specific entity instances. If there is no :id in the template, then ':id'
+ *   is automatically added to its end. Custom methods can use other named parts.
+ * @param config Configure the CRUD behavior:
+ *  - `customMethods`: custom request handlers overriding the defaults.
+ *     Each present property overrides that method's handler. You can use `standardCRUDMethods` members here to
+ *     conveniently reject some actions. When `customMethods` is absent, we use methods from `defaultCrudMethods`.
+ *     Note that these default methods look for the `id` property in their `params` argument; if set, its value is
+ *     the id of the entity to process. Conveniently, the default `urlTemplate` parser sets this property from the
+ *     `:id` pattern.
+ *  - `createResponses`: if present, a dictionary of method-specific Response creators.
+ *  - `defaultCreateResponse`: default function to create all responses if `createResponses` entry is not provided.
+ *     Defaults to `responseFromJson()`.
+ *  - `parsePath`: parses the URL path instead of https://deno.land/x/regexparam. The parsing result is passed to
+ *     CRUD methods as the `params` argument.
+ * @returns A request-handling function suitable as a default export in an endpoint.
+ */
+export function crud<
+    T extends ChiselEntity,
+    E extends ChiselEntityClass<T>,
+    P extends CRUDBaseParams = CRUDBaseParams,
+>(
+    entity: E,
+    urlTemplateSuffix: string,
+    config?: {
+        customMethods?: Partial<CRUDMethods<T, ChiselEntityClass<T>, P>>;
+        createResponses?: Partial<
+            CRUDCreateResponses<T, ChiselEntityClass<T>, P>
+        >;
+        defaultCreateResponse?: CRUDCreateResponse;
+        parsePath?: (url: URL) => P;
+    },
+): (req: Request) => Promise<Response> {
+    const pathTemplateRaw = "/:chiselVersion" + requestContext.path + "/" +
+        (urlTemplateSuffix.includes(":id")
+            ? urlTemplateSuffix
+            : `${urlTemplateSuffix}/:id`);
+
+    const pathTemplate = pathTemplateRaw.replace(/\/+/g, "/"); // in case we end up with foo///bar somehow.
+
+    const defaultCreateResponse = config?.defaultCreateResponse ||
+        responseFromJson;
+    const parsePath = config?.parsePath ||
+        createURLPathParser(pathTemplate);
+    const localDefaultCrudMethods =
+        defaultCrudMethods as unknown as CRUDMethods<T, E, P>;
+    const methods = config?.customMethods
+        ? { ...localDefaultCrudMethods, ...config?.customMethods }
+        : localDefaultCrudMethods;
+
+    return (req: Request): Promise<Response> => {
+        const methodName = req.method as keyof typeof methods; // assume valid, will be handled gracefully
+        const createResponse = config?.createResponses?.[methodName] ||
+            defaultCreateResponse;
+        const method = methods[methodName];
+        if (!method) {
+            return Promise.resolve(
+                createResponse(`Unsupported HTTP method: ${methodName}`, 405),
+            );
+        }
+
+        const url = new URL(req.url);
+        const params = parsePath(url);
+        return method(entity, req, params, url, createResponse);
+    };
+}
