+++ api/src/datastore.ts
@@ -1,47 +1,7 @@
-// SPDX-FileCopyrightText: © 2021 ChiselStrike <info@chiselstrike.com>
+// SPDX-FileCopyrightText: © 2022 ChiselStrike <info@chiselstrike.com>
 
-function opSync(opName: string, a?: unknown, b?: unknown): unknown {
-    return Deno.core.opSync(opName, a, b);
-}
-
-function opAsync(opName: string, a?: unknown, b?: unknown): Promise<unknown> {
-    return Deno.core.opAsync(opName, a, b);
-}
-
-/**
- * Acts the same as Object.assign, but performs deep merge instead of a shallow one.
- */
-function mergeDeep(
-    target: Record<string, unknown>,
-    ...sources: Record<string, unknown>[]
-): Record<string, unknown> {
-    function isObject(item: unknown): boolean {
-        return (item && typeof item === "object" &&
-            !Array.isArray(item)) as boolean;
-    }
-
-    if (!sources.length) {
-        return target;
-    }
-    const source = sources.shift();
-
-    if (isObject(target) && isObject(source)) {
-        for (const key in source) {
-            if (isObject(source[key])) {
-                if (!target[key]) {
-                    Object.assign(target, { [key]: {} });
-                }
-                mergeDeep(
-                    target[key] as Record<string, unknown>,
-                    source[key] as Record<string, unknown>,
-                );
-            } else {
-                Object.assign(target, { [key]: source[key] });
-            }
-        }
-    }
-    return mergeDeep(target, ...sources);
-}
+import { crud } from "./crud.ts";
+import { mergeDeep, opAsync, opSync } from "./utils.ts";
 
 /**
  * Base class for various Operators applicable on `ChiselCursor`. An
@@ -694,149 +654,6 @@ export class ChiselCursor<T> {
     }
 }
 
-/** Extends the Request class adding ChiselStrike-specific helpers
- *
- * @property {string} version - The current API Version
- * @property {string} endpoint - The current endpoint being called.
- * @property {string} pathParams - This is essentially the URL's path, but with everything before the endpoint name removed.
- * @property {AuthUser} user - The currently logged in user. `undefined` if there isn't one.
- * @property {Query} query - Helper structure containing parsed query string from the URL.
- */
-export class ChiselRequest extends Request {
-    public query: Query;
-
-    constructor(
-        input: string,
-        init: RequestInit,
-        public version: string,
-        public endpoint: string,
-        public pathParams: string,
-        public user?: AuthUser | undefined,
-    ) {
-        super(input, init);
-        this.query = new Query(new URL(this.url).searchParams);
-    }
-
-    /**
-     * Returns each component of the arguments part of the path
-     *
-     * While you could call split() on pathParams directly, this
-     * convenience function is useful as it handle empty strings better.
-     *
-     * For example, for the endpoint `/dev/name` this will return an empty
-     * array, while pathParams will be "", and splitting that by "/" returns an
-     * array with one element, the empty string
-     */
-    pathComponents(): string[] {
-        return this.pathParams.split("/").filter((n) => n.length != 0);
-    }
-}
-
-/** Query is a helper class used to access query parameters parsed from the URL.
- */
-export class Query {
-    constructor(private params: URLSearchParams) {}
-
-    /**
-     * Gets the first query parameter named `paramName`. If no such a parameter exists, returns
-     * `undefined`.
-     * @param paramName query parameter to be retrieved from the URL's query string.
-     */
-    get(paramName: string): string | undefined {
-        return this.params.get(paramName) ?? undefined;
-    }
-
-    /**
-     * Gets the first query parameter named `paramName` and tries to parse it as number. If no such a
-     * parameter exists or the parsing fails, returns `undefined`.
-     * @param paramName query parameter to be retrieved from the URL's query string.
-     */
-    getNumber(paramName: string): number | undefined {
-        const v = this.get(paramName);
-        if (v !== undefined) {
-            const f = Number.parseFloat(v);
-            if (Number.isNaN(f)) {
-                return undefined;
-            } else {
-                return f;
-            }
-        }
-        return undefined;
-    }
-
-    /**
-     * Gets the first query parameter named `paramName` and tries to parse it as boolean. If no such a
-     * parameter exists, returns `undefined`
-     * If `paramName` key exists, the value is first converted to lowercase and then 'false', '' and
-     * '0' are evaluated as false, anything else as true.
-     * @param paramName query parameter to be retrieved from the URL's query string.
-     */
-    getBool(paramName: string): boolean | undefined {
-        const v = this.get(paramName);
-        if (v !== undefined) {
-            switch (v.toLowerCase()) {
-                case "false":
-                case "0":
-                case "":
-                    return false;
-                default:
-                    return true;
-            }
-        }
-        return undefined;
-    }
-
-    /**
-     * Returns all the values association with a given query parameter.
-     */
-    getAll(name: string): string[] {
-        return this.params.getAll(name);
-    }
-
-    /**
-     * Returns a Boolean indicating if such a query parameter exists.
-     */
-    has(name: string): boolean {
-        return this.params.has(name);
-    }
-
-    /**
-     * The entries() method returns an iterator allowing iteration through all
-     * key/value pairs contained in the Query. The iterator returns key/value
-     * pairs in the same order as they appear in the query string.
-     */
-    entries(): IterableIterator<[string, string]> {
-        return this.params.entries();
-    }
-
-    /**
-     * The keys() method returns an iterator allowing iteration through all
-     * keys contained in the Query. The iterator returns keys
-     * in the same order as they appear in the query string.
-     */
-    keys(): IterableIterator<string> {
-        return this.params.keys();
-    }
-
-    /**
-     * The values() method returns an iterator allowing iteration through all
-     * values contained in the Query. The iterator returns values
-     * in the same order as they appear in the query string.
-     */
-    values(): IterableIterator<string> {
-        return this.params.values();
-    }
-
-    /**
-     * The Query object provides an iterator equivalent to the iterator provided
-     * by the `entries()` method. For further documentation, please see the docs of
-     * `entries()` method.
-     */
-    [Symbol.iterator](): IterableIterator<[string, string]> {
-        return this.entries();
-    }
-}
-
 export function chiselIterator<T extends ChiselEntity>(
     type: { new (): T },
 ) {
@@ -1224,47 +1041,6 @@ function restrictionsToFilterExpr<T extends ChiselEntity>(
     return expr;
 }
 
-export class AuthUser extends ChiselEntity {
-    emailVerified?: string;
-    name?: string;
-    email?: string;
-    image?: string;
-}
-
-/**
- * Gets a secret from the environment
- *
- * To allow a secret to be used, the server has to be run with * --allow-env <YOUR_SECRET>
- *
- * In development mode, all of your environment variables are accessible
- */
-type JSONValue =
-    | string
-    | number
-    | boolean
-    | null
-    | { [x: string]: JSONValue }
-    | Array<JSONValue>;
-
-export function getSecret(key: string): JSONValue | undefined {
-    return opSync("op_chisel_get_secret", key) as JSONValue | undefined;
-}
-
-export function responseFromJson(body: unknown, status = 200) {
-    // https://fetch.spec.whatwg.org/#null-body-status
-    const isNullBody = (status: number): boolean => {
-        return status == 101 || status == 204 || status == 205 || status == 304;
-    };
-
-    const json = isNullBody(status) ? null : JSON.stringify(body, null, 2);
-    return new Response(json, {
-        status: status,
-        headers: [
-            ["content-type", "application/json"],
-        ],
-    });
-}
-
 export function labels(..._val: string[]) {
     return <T>(_target: T, _propertyName: string) => {
         // chisel-decorator, no content
@@ -1273,21 +1049,6 @@ export function labels(..._val: string[]) {
 
 export function unique(_target: unknown, _name: string): void {
     // chisel-decorator, no content
-}
-
-/** Returns the currently logged-in user or null if no one is logged in. */
-export async function loggedInUser(): Promise<AuthUser | undefined> {
-    const id = requestContext.userId;
-    if (id === undefined) {
-        return undefined;
-    }
-    return await AuthUser.findOne({ id });
-}
-
-function ensureNotGet() {
-    if (requestContext.method === "GET") {
-        throw new Error("Mutating the backend is not allowed during GET");
-    }
 }
 
 export const requestContext: {
@@ -1303,432 +1064,24 @@ export const requestContext: {
     apiVersion: "",
 };
 
-// TODO: BEGIN: this should be in another file: crud.ts
-
-// TODO: BEGIN: when module import is fixed:
-//     import { parse as regExParamParse } from "regexparam";
-// or:
-//     import { parse as regExParamParse } from "regexparam";
-// In the meantime, the regExParamParse function is copied from
-// https://deno.land/x/regexparam@v2.0.0/src/index.js under MIT License included
-// below. ChiselStrike added the TS signature and minor cleanups.
-//
-// Copyright (c) Luke Edwards <luke.edwards05@gmail.com> (lukeed.com)
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-export function regExParamParse(str: string, loose: boolean) {
-    let tmp, pattern = "";
-    const keys = [], arr = str.split("/");
-    arr[0] || arr.shift();
-
-    while ((tmp = arr.shift())) {
-        const c = tmp[0];
-        if (c === "*") {
-            keys.push("wild");
-            pattern += "/(.*)";
-        } else if (c === ":") {
-            const o = tmp.indexOf("?", 1);
-            const ext = tmp.indexOf(".", 1);
-            keys.push(tmp.substring(1, ~o ? o : ~ext ? ext : tmp.length));
-            pattern += !!~o && !~ext ? "(?:/([^/]+?))?" : "/([^/]+?)";
-            if (~ext) pattern += (~o ? "?" : "") + "\\" + tmp.substring(ext);
-        } else {
-            pattern += "/" + tmp;
-        }
+function ensureNotGet() {
+    if (requestContext.method === "GET") {
+        throw new Error("Mutating the backend is not allowed during GET");
     }
-
-    return {
-        keys: keys,
-        pattern: new RegExp("^" + pattern + (loose ? "(?=$|/)" : "/?$"), "i"),
-    };
 }
-// TODO: END: when module import is fixed
 
-type ChiselEntityClass<T extends ChiselEntity> = {
-    new (): T;
-    findOne: (_: { id: string }) => Promise<T | undefined>;
-    findMany: (_: Partial<T>) => Promise<T[]>;
-    build: (...properties: Record<string, unknown>[]) => T;
-    delete: (restrictions: Partial<T>) => Promise<void>;
-    cursor: () => ChiselCursor<T>;
-};
+export class AuthUser extends ChiselEntity {
+    emailVerified?: string;
+    name?: string;
+    email?: string;
+    image?: string;
+}
 
-type GenericChiselEntityClass = ChiselEntityClass<ChiselEntity>;
-
-/**
- * Creates a path parser from a template using regexparam.
- *
- * @param pathTemplate the path template such as `/static`, `/param/:id/:otherParam`...
- * @param loose if true, it can match longer paths. False by default
- * @returns function that can parse paths given as string.
- * @see https://deno.land/x/regexparam@v2.0.0
- */
-export function createPathParser<T extends Record<string, unknown>>(
-    pathTemplate: string,
-    loose = false,
-): (path: string) => T {
-    const { pattern, keys: keysOrFalse } = regExParamParse(pathTemplate, loose);
-    if (typeof keysOrFalse === "boolean") {
-        throw new Error(
-            `invalid pathTemplate=${pathTemplate}, expected string`,
-        );
+/** Returns the currently logged-in user or null if no one is logged in. */
+export async function loggedInUser(): Promise<AuthUser | undefined> {
+    const id = requestContext.userId;
+    if (id === undefined) {
+        return undefined;
     }
-    const keys = keysOrFalse;
-    return function pathParser(path: string): T {
-        const matches = pattern.exec(path);
-        return keys.reduce(
-            (acc: Record<string, unknown>, key: string, index: number) => {
-                acc[key] = matches?.[index + 1];
-                return acc;
-            },
-            {},
-        ) as T;
-    };
+    return await AuthUser.findOne({ id });
 }
-
-/**
- * Creates a path parser from a template using regexparam.
- *
- * @param pathTemplate the path template such as `/static`, `/param/:id/:otherParam`...
- * @param loose if true, it can match longer paths. False by default
- * @returns function that can parse paths given in URL.pathname.
- * @see https://deno.land/x/regexparam@v2.0.0
- */
-export function createURLPathParser<T extends Record<string, unknown>>(
-    pathTemplate: string,
-    loose = false,
-): (url: URL) => T {
-    const pathParser = createPathParser<T>(pathTemplate, loose);
-    return (url: URL): T => pathParser(url.pathname);
-}
-
-/** Creates a Response object from response body and status. */
-export type CRUDCreateResponse = (
-    body: unknown,
-    status: number,
-) => Promise<Response> | Response;
-
-export type CRUDBaseParams = {
-    /** identifier of the object being manipulated, if any */
-    id?: string;
-    /** ChiselStrike's version/branch the server is running,
-     * such as 'dev' for endpoint '/dev/example'
-     * when using 'chisel apply --version dev'
-     */
-    chiselVersion: string;
-};
-
-export type CRUDMethodSignature<
-    T extends ChiselEntity,
-    E extends ChiselEntityClass<T>,
-    P extends CRUDBaseParams = CRUDBaseParams,
-> = (
-    entity: E,
-    req: Request,
-    params: P,
-    url: URL,
-    createResponse: CRUDCreateResponse,
-) => Promise<Response>;
-
-/**
- * A dictionary mapping HTTP methods into corresponding REST methods that process a Request and return a Response.
- */
-export type CRUDMethods<
-    T extends ChiselEntity,
-    E extends ChiselEntityClass<T>,
-    P extends CRUDBaseParams = CRUDBaseParams,
-> = {
-    GET: CRUDMethodSignature<T, E, P>;
-    POST: CRUDMethodSignature<T, E, P>;
-    PUT: CRUDMethodSignature<T, E, P>;
-    PATCH: CRUDMethodSignature<T, E, P>;
-    DELETE: CRUDMethodSignature<T, E, P>;
-};
-
-export type CRUDCreateResponses<
-    T extends ChiselEntity,
-    E extends ChiselEntityClass<T>,
-    P extends CRUDBaseParams = CRUDBaseParams,
-> = {
-    [K in keyof CRUDMethods<T, E, P>]: CRUDCreateResponse;
-};
-
-/**
- * Fetches crud data based on crud `url`.
- */
-async function fetchEntitiesCrud<T extends ChiselEntity>(
-    type: { new (): T },
-    url: string,
-): Promise<T[]> {
-    const results = await opAsync(
-        "op_chisel_crud_query",
-        {
-            typeName: type.name,
-            url,
-        },
-        requestContext,
-    );
-    return results as T[];
-}
-
-async function deleteEntitiesCrud<T extends ChiselEntity>(
-    type: { new (): T },
-    url: string,
-): Promise<void> {
-    await opAsync(
-        "op_chisel_crud_delete",
-        {
-            typeName: type.name,
-            url,
-        },
-        requestContext,
-    );
-}
-
-const defaultCrudMethods: CRUDMethods<ChiselEntity, GenericChiselEntityClass> =
-    {
-        // Returns a specific entity matching params.id (if present) or all entities matching the filter in the `filter` URL parameter.
-        GET: async (
-            entity: GenericChiselEntityClass,
-            _req: Request,
-            params: CRUDBaseParams,
-            url: URL,
-            createResponse: CRUDCreateResponse,
-        ) => {
-            const { id } = params;
-            if (id) {
-                const u = await entity.findOne({ id });
-                return createResponse(u ?? "Not found", u ? 200 : 404);
-            } else {
-                return createResponse(
-                    await fetchEntitiesCrud(entity, url.href),
-                    200,
-                );
-            }
-        },
-        // Creates and returns a new entity from the `req` payload. Ignores the payload's id property and assigns a fresh one.
-        POST: async (
-            entity: GenericChiselEntityClass,
-            req: Request,
-            _params: CRUDBaseParams,
-            _url: URL,
-            createResponse: CRUDCreateResponse,
-        ) => {
-            const u = entity.build(await req.json());
-            u.id = undefined;
-            await u.save();
-            return createResponse(u, 200);
-        },
-        // Updates and returns the entity matching params.id (which must be set) from the `req` payload.
-        PUT: async (
-            entity: GenericChiselEntityClass,
-            req: Request,
-            params: CRUDBaseParams,
-            _url: URL,
-            createResponse: CRUDCreateResponse,
-        ) => {
-            const { id } = params;
-            if (!id) {
-                return createResponse(
-                    "PUT requires item ID in the URL",
-                    400,
-                );
-            }
-            const u = entity.build(await req.json());
-            u.id = id;
-            await u.save();
-            return createResponse(u, 200);
-        },
-        PATCH: async (
-            entity: GenericChiselEntityClass,
-            req: Request,
-            params: CRUDBaseParams,
-            _url: URL,
-            createResponse: CRUDCreateResponse,
-        ) => {
-            const { id } = params;
-            if (!id) {
-                return createResponse(
-                    "PATCH requires item ID in the URL",
-                    400,
-                );
-            }
-            const orig = await entity.findOne({ id });
-            if (!orig) {
-                return createResponse(
-                    "object does not exist, cannot PATCH",
-                    404,
-                );
-            }
-            mergeDeep(
-                orig as unknown as Record<string, unknown>,
-                await req.json(),
-            );
-            await orig.save();
-            return createResponse(orig, 200);
-        },
-
-        // Deletes the entity matching params.id (if present) or all entities matching the filter in the `filter` URL parameter. One of the two must be present.
-        DELETE: async (
-            entity: GenericChiselEntityClass,
-            _req: Request,
-            params: CRUDBaseParams,
-            url: URL,
-            createResponse: CRUDCreateResponse,
-        ) => {
-            const { id } = params;
-            if (id) {
-                await entity.delete({ id });
-                return createResponse(`Deleted ID ${id}`, 200);
-            } else {
-                await deleteEntitiesCrud(entity, url.href);
-                return createResponse(
-                    `Deleted entities matching ${url.search}`,
-                    200,
-                );
-            }
-        },
-    } as const;
-
-/**
- * These methods can be used as `customMethods` in `ChiselStrike.crud()`.
- *
- * @example
- * Put this in the file 'endpoints/comments.ts':
- * ```typescript
- * import { Comment } from "../models/comment";
- * export default crud(
- *   Comment,
- *   ':id',
- *   {
- *     PUT: standardCRUDMethods.notFound, // do not update, instead returns 404
- *     DELETE: standardCRUDMethods.methodNotAllowed, // do not delete, instead returns 405
- *   },
- * );
- * ```
- */
-export const standardCRUDMethods = {
-    forbidden: (
-        _entity: GenericChiselEntityClass,
-        _req: Request,
-        _params: CRUDBaseParams,
-        _url: URL,
-        createResponse: CRUDCreateResponse,
-    ) => Promise.resolve(createResponse("Forbidden", 403)),
-    notFound: (
-        _entity: GenericChiselEntityClass,
-        _req: Request,
-        _params: CRUDBaseParams,
-        _url: URL,
-        createResponse: CRUDCreateResponse,
-    ) => Promise.resolve(createResponse("Not Found", 404)),
-    methodNotAllowed: (
-        _entity: GenericChiselEntityClass,
-        _req: Request,
-        _params: CRUDBaseParams,
-        _url: URL,
-        createResponse: CRUDCreateResponse,
-    ) => Promise.resolve(createResponse("Method Not Allowed", 405)),
-} as const;
-
-/**
- * Generates endpoint code to handle REST methods GET/PUT/POST/DELETE for this entity.
- * @example
- * Put this in the file 'endpoints/comments.ts':
- * ```typescript
- * import { Comment } from "../models/comment";
- * export default crud(Comment, ":id");
- * ```
- * This results in a /comments endpoint that correctly handles all REST methods over Comment.
- * @param entity Entity type
- * @param urlTemplateSuffix A suffix to be added to the Request URL (see https://deno.land/x/regexparam for syntax).
- *   Some CRUD methods rely on parts of the URL to identify the resource to apply to. Eg, GET /comments/1234
- *   returns the comment entity with id=1234, while GET /comments returns all comments. This parameter describes
- *   how to find the relevant parts in the URL. Default CRUD methods (see `defaultCrudMethods`) look for the :id
- *   part in this template to identify specific entity instances. If there is no :id in the template, then ':id'
- *   is automatically added to its end. Custom methods can use other named parts.
- * @param config Configure the CRUD behavior:
- *  - `customMethods`: custom request handlers overriding the defaults.
- *     Each present property overrides that method's handler. You can use `standardCRUDMethods` members here to
- *     conveniently reject some actions. When `customMethods` is absent, we use methods from `defaultCrudMethods`.
- *     Note that these default methods look for the `id` property in their `params` argument; if set, its value is
- *     the id of the entity to process. Conveniently, the default `urlTemplate` parser sets this property from the
- *     `:id` pattern.
- *  - `createResponses`: if present, a dictionary of method-specific Response creators.
- *  - `defaultCreateResponse`: default function to create all responses if `createResponses` entry is not provided.
- *     Defaults to `responseFromJson()`.
- *  - `parsePath`: parses the URL path instead of https://deno.land/x/regexparam. The parsing result is passed to
- *     CRUD methods as the `params` argument.
- * @returns A request-handling function suitable as a default export in an endpoint.
- */
-export function crud<
-    T extends ChiselEntity,
-    E extends ChiselEntityClass<T>,
-    P extends CRUDBaseParams = CRUDBaseParams,
->(
-    entity: E,
-    urlTemplateSuffix: string,
-    config?: {
-        customMethods?: Partial<CRUDMethods<T, ChiselEntityClass<T>, P>>;
-        createResponses?: Partial<
-            CRUDCreateResponses<T, ChiselEntityClass<T>, P>
-        >;
-        defaultCreateResponse?: CRUDCreateResponse;
-        parsePath?: (url: URL) => P;
-    },
-): (req: Request) => Promise<Response> {
-    const pathTemplateRaw = "/:chiselVersion" + requestContext.path + "/" +
-        (urlTemplateSuffix.includes(":id")
-            ? urlTemplateSuffix
-            : `${urlTemplateSuffix}/:id`);
-
-    const pathTemplate = pathTemplateRaw.replace(/\/+/g, "/"); // in case we end up with foo///bar somehow.
-
-    const defaultCreateResponse = config?.defaultCreateResponse ||
-        responseFromJson;
-    const parsePath = config?.parsePath ||
-        createURLPathParser(pathTemplate);
-    const localDefaultCrudMethods =
-        defaultCrudMethods as unknown as CRUDMethods<T, E, P>;
-    const methods = config?.customMethods
-        ? { ...localDefaultCrudMethods, ...config?.customMethods }
-        : localDefaultCrudMethods;
-
-    return (req: Request): Promise<Response> => {
-        const methodName = req.method as keyof typeof methods; // assume valid, will be handled gracefully
-        const createResponse = config?.createResponses?.[methodName] ||
-            defaultCreateResponse;
-        const method = methods[methodName];
-        if (!method) {
-            return Promise.resolve(
-                createResponse(`Unsupported HTTP method: ${methodName}`, 405),
-            );
-        }
-
-        const url = new URL(req.url);
-        const params = parsePath(url);
-        return method(entity, req, params, url, createResponse);
-    };
-}
-// TODO: END: this should be in another file: crud.ts
-
-export type ChiselEvent = {
-    key: ArrayBuffer;
-    value: ArrayBuffer;
-};
