+++ api/src/event.ts
@@ -1,0 +1,4 @@
+export type ChiselEvent = {
+    key: ArrayBuffer;
+    value: ArrayBuffer;
+};
