+++ endpoint_tsc/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 tsc_compile = { path = "../tsc_compile" }
 anyhow = "1.0"
 api = { path = "../api" }
-tempfile = "3.2.0"
 
 [lib]
 name = "endpoint_tsc"
