+++ packages/build.rs
@@ -1,11 +1,51 @@
-use api::chisel_d_ts;
-use api::chisel_js;
+#![allow(clippy::expect_fun_call)]
 use std::env;
 use std::fs;
 use std::path::{Path, PathBuf};
 use std::process::Command;
 
-fn run_in<T: IntoIterator<Item = &'static str>>(cmd: &str, args: T, dir: PathBuf) {
+fn main() {
+    build_chiselstrike_api();
+    build_create_chiselstrike_app();
+}
+
+// Build the package `@chiselstrike/api`
+fn build_chiselstrike_api() {
+    let package_dir = Path::new("./chiselstrike-api").to_path_buf();
+
+    // cleanup the `lib/` directory
+    let lib_dir = package_dir.join("lib");
+    mkdir_empty(&lib_dir);
+
+    for (name, code) in api::SOURCES_D_TS.iter() {
+        let output_path = lib_dir.join(format!("{}.ts.d.ts", name));
+        write(&output_path, code);
+    }
+
+    println!("cargo:rerun-if-changed=./chiselstrike-api/package.json");
+}
+
+// Build the package `create-chiselstrike-app`
+fn build_create_chiselstrike_app() {
+    let create_app = Path::new("./create-chiselstrike-app").to_path_buf();
+    // build create-chiselstrike-app so we can use it in tests
+    for v in [
+        "create-chiselstrike-app/index.ts",
+        "create-chiselstrike-app/template/Chisel.toml",
+        "create-chiselstrike-app/template/hello.ts",
+        "create-chiselstrike-app/template/package.json",
+        "create-chiselstrike-app/template/settings.json",
+        "create-chiselstrike-app/template/tsconfig.json",
+        "create-chiselstrike-app/tsconfig.json",
+        "create-chiselstrike-app/package.json",
+    ] {
+        println!("cargo:rerun-if-changed=./{}", v);
+    }
+    run_in("npm", &["install"], create_app.clone());
+    run_in("npm", &["run", "build"], create_app);
+}
+
+fn run_in(cmd: &str, args: &[&str], dir: PathBuf) {
     assert!(
         dir.exists(),
         "{:?} does not exist. Current directory is {:?}",
@@ -27,30 +67,13 @@ fn run_in<T: IntoIterator<Item = &'static str>>(cmd: &str, args: T, dir: PathBuf
     assert!(status.unwrap().success());
 }
 
-fn main() {
-    let create_app = Path::new("./create-chiselstrike-app").to_path_buf();
-    let api = Path::new("./chiselstrike-api").to_path_buf();
-
-    fs::write("./chiselstrike-api/lib/chisel.js", chisel_js()).unwrap();
-    fs::write("./chiselstrike-api/lib/chisel.d.ts", chisel_d_ts()).unwrap();
-
-    // build create-chiselstrike-app so we can use it in tests
-    for v in [
-        "create-chiselstrike-app/index.ts",
-        "create-chiselstrike-app/template/Chisel.toml",
-        "create-chiselstrike-app/template/hello.ts",
-        "create-chiselstrike-app/template/package.json",
-        "create-chiselstrike-app/template/settings.json",
-        "create-chiselstrike-app/template/tsconfig.json",
-        "create-chiselstrike-app/tsconfig.json",
-        "create-chiselstrike-app/package.json",
-    ] {
-        println!("cargo:rerun-if-changed=./{}", v);
+fn mkdir_empty(dir: &Path) {
+    if dir.exists() {
+        fs::remove_dir_all(&dir).expect(&format!("Cannot cleanup directory {}", dir.display()));
     }
-    run_in("npm", ["install"], create_app.clone());
-    run_in("npm", ["run", "build"], create_app);
+    fs::create_dir_all(&dir).expect(&format!("Cannot create directory {}", dir.display()));
+}
 
-    println!("cargo:rerun-if-changed=./chiselstrike-api/package.json");
-    run_in("npm", ["install"], api.clone());
-    run_in("npm", ["run", "build"], api);
+fn write(output: &Path, data: &str) {
+    fs::write(output, data).expect(&format!("Cannot write to file {}", output.display()));
 }
