--- packages/chiselstrike-api/lib/lib.deno_core.d.ts
@@ -1,1 +1,0 @@
-../../third_party/deno/core/lib.deno_core.d.ts
