--- packages/chiselstrike-api/src/chisel.ts
@@ -1,1 +1,0 @@
-../../../api/src/chisel.ts
