--- packages/chiselstrike-api/src/lib.deno_core.d.ts
@@ -1,1 +1,0 @@
-../../../api/src/lib.deno_core.d.ts
